[search] utilize GroupTagValue.last_seen when filtering results
This improves search behavior around high frequency tags. For example, if you had a tag that was on server name, its likely every issue ever has been seen on that tag value. Because of that, you often wouldn't receive the results you're looking for in more complex filters as the first 1,000 results returned were old/or not matching those filters. This changes it to sort them by the last_seen value, which will generally ensure the results you're expecting are visible.